+++ src/js/cell_plugins/cell_plugin_mixin.js
@@ -1,0 +1,23 @@
+'use strict';
+
+/**
+*   @param {domElement} el the plugin target element reference
+*   @param {object} options
+*   @param {object} options.queryState: a copy of the table's current query state
+*   @param {object} options.record: this table row's record object
+*   @param {object} options.event: the click event that triggered plugin construction
+*/
+function CellPluginMixin(el, options) {
+    this._$el = $(el);
+    this._initialQueryState = options.queryState;
+    this._record = options.record;
+    return this;
+}
+
+if (typeof module === 'object') {
+  module.exports = CellPluginMixin;
+} else if (typeof define === 'function') {
+  define(function() { return CellPluginMixin; });
+} else {
+  window.CellPluginMixin = CellPluginMixin;
+}